+++ src/elements/conditions/customers/DiscountCustomerCondition.php
@@ -19,7 +19,7 @@ class DiscountCustomerCondition extends UserElementCondition
     {
         return array_merge(parent::conditionRuleTypes(), [
             HasOrdersConditionRule::class,
-            LoggedInConditionRule::class,
+            SignedInConditionRule::class,
         ]);
     }
 }